chore(test-file): add shortAnswer test file
